Adopt business layer to Qt6

+++ src/corelib/business_layer/model/characters/characters_model.cpp
@@ -372,8 +372,9 @@ void CharactersModel::initDocument()
         //
         for (int index = 0; index < d->characterModels.size(); ++index) {
             if (d->characterModels.at(index)->name() == characterName) {
-                d->characterModels.move(index,
-                                        std::min(characterIndex++, d->characterModels.size() - 1));
+                d->characterModels.move(
+                    index,
+                    std::min(characterIndex++, static_cast<int>(d->characterModels.size() - 1)));
             }
         }
         //
