+++ src/corelib/business_layer/plots/audioplay/audioplay_structure_analysis_plot.cpp
@@ -406,7 +406,8 @@ void AudioplayStructureAnalysisPlot::saveToFile(const QString& _fileName) const
             }
             ++reportRow;
 
-            firstColumnWidth = std::max(firstColumnWidth, sceneInfo->constFirst().size());
+            firstColumnWidth
+                = std::max(firstColumnWidth, static_cast<int>(sceneInfo->constFirst().size()));
         }
     }
